refactor: Create single entrypoint for invoice utils
  - Remove CLI entrypoint
  - Make web endpoints the main entrypoint
