+++ src/invoice_utils/api/__init__.py
@@ -1,5 +1,7 @@
 from ._template import router as template_router
 from ._templates import router as templates_router
+from ._errors import InvoiceRequestInputError, InvoiceRequestEmailError, input_error_handler, email_error_handler
+from ._invoices import router as invoices_router
 
-
-__all__ = ["template_router", "templates_router"]
+__all__ = ["template_router", "templates_router", "InvoiceRequestEmailError", "InvoiceRequestInputError",
+           "input_error_handler", "email_error_handler", "invoices_router"]
