+++ tests/invoice_utils/test_web.py
@@ -2,12 +2,13 @@ import smtplib
 from unittest.mock import MagicMock, call, ANY, patch
 
 import pytest
-from fastapi import HTTPException
 from fastapi.testclient import TestClient
 from jinja2 import TemplateNotFound
 
 from invoice_utils.config import DEFAULT_MAIL_SUBJECT, DEFAULT_BODY_TEMPLATE_NAME, DEFAULT_BODY_TEMPLATE_PACKAGE, \
     DEFAULT_TEMPLATES_DIRECTORY, DEFAULT_INVOICE_DIR, DEFAULT_RULE_TEMPLATE_NAME
+
+CREATE_INVOICE_PATH = "/api/v1/invoices"
 
 MESSAGE_ARGUMENT = 2
 
@@ -35,8 +36,8 @@ def email_invoice_request_body(invoice_request_body: dict):
 
 @pytest.fixture
 def mock_smtp(mocker):
-    result = mocker.MagicMock(name="invoice_utils.web.smtplib.SMTP")
-    mocker.patch("invoice_utils.web.smtplib.SMTP", new=result)
+    result = mocker.MagicMock(name="invoice_utils.api._invoices.smtplib.SMTP")
+    mocker.patch("invoice_utils.api._invoices.smtplib.SMTP", new=result)
     return result
 
 
@@ -55,15 +56,15 @@ def expected_body():
 
 @pytest.fixture
 def mock_render(mocker):
-    result = MagicMock(name="invoice_utils.web.PdfInvoiceRenderer.render")
-    mocker.patch("invoice_utils.web.PdfInvoiceRenderer.render", new=result)
+    result = MagicMock(name="invoice_utils.api._invoices.PdfInvoiceRenderer.render")
+    mocker.patch("invoice_utils.api._invoices.PdfInvoiceRenderer.render", new=result)
     result.return_value = b"test pdf content"
     return result
 
 
 def test_send_mail_param_not_provided(http, caplog, mock_render, invoice_request_body):
     with caplog.at_level("INFO"):
-        res = http.post("/invoice", json=invoice_request_body)
+        res = http.post(CREATE_INVOICE_PATH, json=invoice_request_body)
 
     assert "Report was sent to test@email.com" not in caplog.messages
     assert res.status_code == 201
@@ -74,7 +75,7 @@ def test_send_mail_param_fails_without_address_param(
 ):
     invoice_request_body["send_mail"] = True
     with caplog.at_level("INFO"):
-        res = http.post("/invoice", json=invoice_request_body)
+        res = http.post(CREATE_INVOICE_PATH, json=invoice_request_body)
 
     assert "Report was sent to " not in caplog.messages
     assert (
@@ -88,7 +89,7 @@ def test_send_mail_alternate_flow(
     http, caplog, mock_render, email_invoice_request_body, server
 ):
     with caplog.at_level("INFO"):
-        res = http.post("/invoice", json=email_invoice_request_body)
+        res = http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert server.starttls.call_count == 0
     assert server.login.call_count == 0
@@ -101,7 +102,7 @@ def test_send_mail_fails_for_smtp_exception(http, caplog, mock_render, email_inv
     mock_smtp.side_effect = smtplib.SMTPException
 
     with caplog.at_level("INFO"):
-        res = http.post("/invoice", json=email_invoice_request_body)
+        res = http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert "Report was sent to test@email.com" not in caplog.messages
     assert res.json()["message"] == "There was a problem sending the email."
@@ -119,7 +120,7 @@ def test_send_mail_fails_for_smtp_exception(http, caplog, mock_render, email_inv
 def test_email_sent_with_expected_subject(
     environment, http, server, mock_render, email_invoice_request_body, expected_subject,
 ):
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert server.sendmail.call_args_list == [
         call(ANY, ANY, ANY)
@@ -141,7 +142,7 @@ def test_email_sent_with_expected_subject(
 def test_smtp_login_when_user_and_password_are_specified(
     environment, http, server, mock_render, email_invoice_request_body,
 ):
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert server.login.call_args_list == [call("user", "password")]
 
@@ -162,7 +163,7 @@ def test_smtp_login_when_user_and_password_are_specified(
 def test_smtp_starttls_called_when_env_flag_is_set(
     environment, http, server, mock_render, email_invoice_request_body
 ):
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert server.starttls.call_count == 1
 
@@ -179,7 +180,7 @@ def test_smtp_starttls_called_when_env_flag_is_set(
 def test_smtp_starttls_env_flag_must_be_true_boolean(
     environment, http, server, mock_render, email_invoice_request_body
 ):
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert server.starttls.call_count == 0
 
@@ -202,7 +203,7 @@ def test_smtp_starttls_env_flag_must_be_true_boolean(
 def test_email_body_was_sent_with_expected_body(
     environment, http, server, mock_render, email_invoice_request_body, expected_body
 ):
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     email_content = server.sendmail.call_args.args[MESSAGE_ARGUMENT]
     assert expected_body in email_content
@@ -224,7 +225,7 @@ def test_email_body_was_sent_with_expected_body(
 def test_sendmail_was_called_with_pdf_attachment(
     environment, http, server, mock_render, email_invoice_request_body
 ):
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     email_content = server.sendmail.call_args.args[MESSAGE_ARGUMENT]
     assert "Content-Disposition: attachment; filename=\"20231114-0001-invoice.pdf\"" in email_content
@@ -247,7 +248,7 @@ def test_sendmail_was_called_with_pdf_attachment(
 def test_invoice_generated_within_specified_dir(
     environment, http, server, mock_render, email_invoice_request_body
 ):
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert environment.get("INVOICE_UTILS_INVOICE_DIR") in mock_render.call_args.args[1]
 
@@ -269,7 +270,7 @@ def test_invoice_generated_within_specified_dir(
 def test_invoice_generation_fails_if_directory_does_not_exist(
     environment, http, server, mock_render, email_invoice_request_body
 ):
-    res = http.post("/invoice", json=email_invoice_request_body)
+    res = http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
     assert res.status_code == 507
     assert res.json() == {"detail": "No local storage available for invoices"}
 
@@ -291,9 +292,9 @@ def test_invoice_generation_fails_if_directory_does_not_exist(
 def test_invoice_generation_fails_if_directory_does_not_exist(
     environment, http, server, mocker, email_invoice_request_body
 ):
-    with mocker.patch("invoice_utils.web.os.path.isdir", return_value=True):
-        with mocker.patch("invoice_utils.web.os.access", return_value=False):
-            res = http.post("/invoice", json=email_invoice_request_body)
+    with mocker.patch("invoice_utils.api._invoices.os.path.isdir", return_value=True):
+        with mocker.patch("invoice_utils.api._invoices.os.access", return_value=False):
+            res = http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
     assert res.status_code == 507
     assert res.json() == {"detail": "Insufficient rights to store invoice"}
 
@@ -301,10 +302,10 @@ def test_invoice_generation_fails_if_directory_does_not_exist(
 def test_render_body_is_called_with_default_values(
     http, server, mock_render, email_invoice_request_body, mocker
 ):
-    mock_env = mocker.MagicMock(name="invoice_utils.web.Environment")
-    mocker.patch("invoice_utils.web.Environment", new=mock_env)
+    mock_env = mocker.MagicMock(name="invoice_utils.api._invoices.Environment")
+    mocker.patch("invoice_utils.api._invoices.Environment", new=mock_env)
 
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert mock_env.call_args.kwargs["loader"].package_name == DEFAULT_BODY_TEMPLATE_PACKAGE
     assert mock_env.call_args.kwargs["loader"].package_path == DEFAULT_TEMPLATES_DIRECTORY
@@ -324,7 +325,7 @@ def test_template_related_variables_with_bad_inputs_raise_errors(
     environment, http, server, mock_render, email_invoice_request_body, expected_error
 ):
     with pytest.raises(expected_error):
-        http.post("/invoice", json=email_invoice_request_body)
+        http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
 
 @pytest.mark.parametrize(
@@ -338,11 +339,11 @@ def test_app_lifespan_sets_rule_template_to_default_if_not_exists(
     environment, http, server, mock_render, email_invoice_request_body
 ):
     with patch("dotenv.load_dotenv", MagicMock(name="load_dotenv")):
-        import invoice_utils.web as web
+        import invoice_utils.__main__ as web
 
         assert web.config.INVOICE_UTILS_RULE_TEMPLATE_NAME == environment.get("INVOICE_UTILS_RULE_TEMPLATE_NAME")
         with TestClient(web.app) as client:
-            client.post("/invoice", json=email_invoice_request_body)
+            client.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
         assert web.config.INVOICE_UTILS_RULE_TEMPLATE_NAME == DEFAULT_RULE_TEMPLATE_NAME
 
 
@@ -350,18 +351,18 @@ def test_app_lifespan_raise_exception_if_default_not_exists(
     http, server, mock_render, email_invoice_request_body
 ):
     with patch("dotenv.load_dotenv", MagicMock(name="load_dotenv")):
-        import invoice_utils.web as web
+        import invoice_utils.__main__ as web
 
         web.config.DEFAULT_RULE_TEMPLATE_NAME = "non-exist"
         with pytest.raises(Exception):
             with TestClient(web.app) as client:
-                client.post("/invoice", json=email_invoice_request_body)
+                client.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
 
 def test_generate_invoice_calls_repository_to_get_rules(
     http, server, mock_render, email_invoice_request_body, template_repo
 ):
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert template_repo.get_by_key.call_count == 1
 
@@ -376,7 +377,7 @@ def test_generate_invoice_calls_repository_to_get_rules(
 def test_repo_get_by_key_is_called_with_env_variable(
     environment, http, server, mock_render, email_invoice_request_body, template_repo
 ):
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert template_repo.get_by_key.call_args.args[0] == environment.get("INVOICE_UTILS_RULE_TEMPLATE_NAME")
 
@@ -386,7 +387,7 @@ def test_repo_get_by_key_called_twice_if_given_rule_template_name(
 ):
     email_invoice_request_body["rule_template_name"] = "some_template"
 
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert template_repo.get_by_key.call_count == 2
 
@@ -396,7 +397,7 @@ def test_repo_get_by_key_called_with_given_rule_template_name(
 ):
     email_invoice_request_body["rule_template_name"] = "some_template"
 
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert template_repo.get_by_key.call_args.args[0] == "some_template"
 
@@ -406,7 +407,7 @@ def test_repo_get_by_key_overwrites_rules_if_given_valid_rule_template_name(
 ):
     email_invoice_request_body["rule_template_name"] = "some_template"
 
-    http.post("/invoice", json=email_invoice_request_body)
+    http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert template_repo.get_by_key.call_count == 2
     found, rule_template = template_repo.get_by_key.return_value
@@ -423,7 +424,7 @@ def test_generate_invoice_raises_error_if_given_rule_template_name_is_invalid(
     email_invoice_request_body["rule_template_name"] = "invalid_name"
     template_repo.get_by_key.side_effect = [(True, default_template), (False, None)]
 
-    res = http.post("/invoice", json=email_invoice_request_body)
+    res = http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert res.status_code == 400
     assert res.json() == {"detail": "Rule Template does not exist."}
@@ -433,7 +434,7 @@ def test_get_by_key_raises_error_if_default_rule_template_fails(
     http, server, mock_render, email_invoice_request_body, template_repo
 ):
     template_repo.get_by_key.return_value = (False, None)
-    res = http.post("/invoice", json=email_invoice_request_body)
+    res = http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert res.status_code == 400
     assert res.json() == {"detail": "Rule Template does not exist."}
@@ -445,7 +446,7 @@ def test_generate_invoice_raises_error_if_invalid_default_rules_and_invalid_give
     email_invoice_request_body["rule_template_name"] = "invalid_name"
     template_repo.get_by_key.side_effect = [(False, None), (False, None)]
 
-    res = http.post("/invoice", json=email_invoice_request_body)
+    res = http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert res.status_code == 400
     assert res.json() == {"detail": "Rule Template does not exist."}
@@ -456,6 +457,6 @@ def test_generate_invoice_returns_2xx_with_invalid_default_rules_but_valid_given
 ):
     email_invoice_request_body["rule_template_name"] = "valid-name"
     template_repo.get_by_key.side_effect = [(False, None), (True, default_template)]
-    res = http.post("/invoice", json=email_invoice_request_body)
+    res = http.post(CREATE_INVOICE_PATH, json=email_invoice_request_body)
 
     assert res.status_code == 201
